lesson delete complete, bootstrap css(lessonPage, header, messageBoxPage, modal)
