+++ fin/src/main/java/edu/kh/fin/band/lesson/model/dao/LessonDAO.java
@@ -77,6 +77,7 @@ public class LessonDAO {
 	 * @return
 	 */
 	public int deleteLesson(int lessonNo) {
+		
 		return sqlSession.update("lessonMapper.deleteLesson", lessonNo);
 	}
