+++ fin/src/main/resources/mapper/lesson-mapper.xml
@@ -29,7 +29,7 @@
 
 	<select id="lessonList" resultMap="lesson_rm">
 	SELECT 
-	u.USER_NICK, li.MOTTO, li.REGION, TO_CHAR(li.WRITING_DATE, 'YY"년"MM"월"DD"일"')AS WRITING_DATE, LESSON_NO
+	u.USER_NICK, li.MOTTO, li.REGION, TO_CHAR(li.WRITING_DATE, 'YY"년 "MM"월 "DD"일"')AS WRITING_DATE, LESSON_NO
 	FROM LESSON_INFO li
 	JOIN "USER" u USING(USER_NO)
 	WHERE LESSON_FL = 'N'
