+++ fin/src/main/webapp/resources/css/memberListModal.css
@@ -37,7 +37,11 @@
     animation-name: slideOutUp;
   }
   
-  .modal-dialog{
+  /* .modal-dialog.{
+    
+  } */
+  
+  #memberListDialog{
     max-width: 650px;
   }
   
@@ -62,4 +66,8 @@
   
   .modal-body.memberModal{
     height: 400px;
+  }
+
+  h1{
+    padding-left: 0;
   }
